services/auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
redis==5.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
